+++ include/tvm/relax/attrs/manipulate.h
@@ -125,30 +125,6 @@ struct TileAttrs : public tvm::AttrsNode<TileAttrs> {
   }
 };  // struct TileAttrs
 
-/*! \brief Attributes used in cumsum operators */
-struct CumsumAttrs : public tvm::AttrsNode<CumsumAttrs> {
-  Optional<Integer> axis;
-  DataType dtype;
-
-  TVM_DECLARE_ATTRS(CumsumAttrs, "relax.attrs.CumsumAttrs") {
-    TVM_ATTR_FIELD(axis).describe(
-        "Axis along which the cumulative sum is computed."
-        "The default (None) is to compute the cumsum over the flattened array.");
-    TVM_ATTR_FIELD(dtype).describe(
-        "Type of the returned array and of the accumulator in which the elements are summed."
-        "If dtype is not specified, it defaults to the dtype of data.");
-  }
-};  // struct CumsumAttrs
-
-/*! \brief Attributes used in einsum operator */
-struct EinsumAttrs : public tvm::AttrsNode<EinsumAttrs> {
-  String subscripts;
-
-  TVM_DECLARE_ATTRS(EinsumAttrs, "relax.attrs.EinsumAttrs") {
-    TVM_ATTR_FIELD(subscripts).describe("The einsum expression string");
-  }
-};  // struct EinsumAttrs
-
 /*! \brief Attributes used in flip operators */
 struct FlipAttrs : public tvm::AttrsNode<FlipAttrs> {
   Integer axis;
