+++ include/tvm/relax/attrs/statistical.h
@@ -42,6 +42,21 @@ struct StatisticalAttrs : public tvm::AttrsNode<StatisticalAttrs> {
   }
 };  // struct StatisticalAttrs
 
+/*! \brief Attributes used in cumsum operators */
+struct CumsumAttrs : public tvm::AttrsNode<CumsumAttrs> {
+  Optional<Integer> axis;
+  DataType dtype;
+
+  TVM_DECLARE_ATTRS(CumsumAttrs, "relax.attrs.CumsumAttrs") {
+    TVM_ATTR_FIELD(axis).describe(
+        "Axis along which the cumulative sum is computed."
+        "The default (None) is to compute the cumsum over the flattened array.");
+    TVM_ATTR_FIELD(dtype).describe(
+        "Type of the returned array and of the accumulator in which the elements are summed."
+        "If dtype is not specified, it defaults to the dtype of data.");
+  }
+};  // struct CumsumAttrs
+
 }  // namespace relax
 }  // namespace tvm
