+++ python/tvm/relax/op/manipulate.py
@@ -17,7 +17,6 @@
 """Manipulation operators."""
 from typing import List, Optional, Tuple, Union, Callable
 
-from tvm import DataType
 from tvm.ir.expr import PrimExpr
 from tvm.tir import IntImm, FloatImm, IndexMap
 
@@ -389,78 +388,6 @@ def tile(data: Expr, repeats: Union[int, Tuple[int], List[int]]) -> Expr:
     if isinstance(repeats, int):
         repeats = [repeats]
     return _ffi_api.tile(data, repeats)  # type: ignore
-
-
-def cumsum(data: Expr, axis: Optional[int] = None, dtype: Optional[Union[str, DataType]] = None):
-    """Numpy style cumsum op. Return the cumulative inclusive sum of the elements along
-    a given axis.
-
-    Parameters
-    ----------
-    data : relax.Expr
-        The input data to the operator.
-
-    axis : Optional[int]
-        Axis along which the cumulative sum is computed. The default (None) is to compute
-        the cumsum over the flattened array.
-
-    dtype : Optional[Union[str, DataType]]
-        Type of the returned array and of the accumulator in which the elements are summed.
-        If dtype is not specified, it defaults to the dtype of data.
-
-    Returns
-    -------
-    result : relax.Expr
-        The result has the same size as data, and the same shape as data if axis is not None.
-        If axis is None, the result is a 1-d array.
-
-    Examples
-    --------
-    .. code-block:: python
-
-        a = [[1, 2, 3], [4, 5, 6]]
-
-        cumsum(a)  # if axis is not provided, cumsum is done over the flattened input.
-        -> [ 1,  3,  6, 10, 15, 21]
-
-        cumsum(a, dtype="float32")
-        -> [  1.,   3.,   6.,  10.,  15.,  21.]
-
-        cumsum(a, axis=0)  # sum over rows for each of the 3 columns
-        -> [[1, 2, 3],
-            [5, 7, 9]]
-
-        cumsum(a, axis=1)
-        -> [[ 1,  3,  6],
-            [ 4,  9, 15]]
-
-        a = [1, 0, 1, 0, 1, 1, 0]  # a is a boolean array
-        cumsum(a, dtype=int32)  # dtype should be provided to get the expected results
-        -> [1, 1, 2, 2, 3, 4, 4]
-    """
-    return _ffi_api.cumsum(data, axis, dtype)  # type: ignore
-
-
-def einsum(operands, subscripts):
-    """Evaluates the Einstein summation convention on data
-
-    Parameters
-    ----------
-    operands : Union(List[relax.Expr], Tuple[relax.Expr])
-        A list of expression.
-
-    subscripts : str
-        The einsum expression string.
-
-    Returns
-    -------
-    result : relax.Expr
-        The output from the einsum op.
-    """
-    if isinstance(operands, (list, tuple)):
-        operands = RxTuple(operands)
-
-    return _ffi_api.einsum(operands, subscripts)  # type: ignore
 
 
 def flip(data, axis):
