+++ python/tvm/relax/transform/legalize_ops/statistical.py
@@ -81,3 +81,8 @@ register_legalize("relax.max", _statistical(topi.max))
 register_legalize("relax.min", _statistical(topi.min))
 register_legalize("relax.prod", _statistical(topi.prod))
 register_legalize("relax.sum", _statistical(topi.sum))
+
+
+@register_legalize("relax.cumsum")
+def _cumsum(bb: BlockBuilder, call: Call) -> Expr:
+    return bb.call_te(topi.cumsum, call.args[0], call.attrs.axis, call.attrs.dtype)
