+++ tests/python/relax/test_transform_legalize_ops_manipulate.py
@@ -1280,174 +1280,6 @@ def test_tile_symbolic():
     tvm.ir.assert_structural_equal(mod, Expected)
 
 
-def test_cumsum():
-    # fmt: off
-    @I.ir_module
-    class Cumsum:
-        @R.function
-        def main(x: R.Tensor((3, 2, 3), "float32")):
-            gv = R.cumsum(x, axis=1, dtype="int32")
-            return gv
-
-    @I.ir_module
-    class Expected:
-        @T.prim_func
-        def cumsum(var_rxplaceholder: T.handle, out_buf: T.Buffer((T.int64(3), T.int64(2), T.int64(3)), "int32")):
-            T.func_attr({"tir.noalias": True})
-            rxplaceholder = T.match_buffer(var_rxplaceholder, (T.int64(3), T.int64(2), T.int64(3)), offset_factor=1)
-            with T.block("cumsum_generic"):
-                T.reads(rxplaceholder[T.int64(0):T.int64(3), T.int64(0):T.int64(2), T.int64(0):T.int64(3)])
-                T.writes(out_buf[T.int64(0):T.int64(3), T.int64(0):T.int64(2), T.int64(0):T.int64(3)])
-                for fused in T.parallel(T.int64(9)):
-                    out_buf[(fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) // T.int64(3) // T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) // T.int64(3) % T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) % T.int64(3)] = T.Cast("int32", rxplaceholder[(fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) // T.int64(3) // T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) // T.int64(3) % T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3)) % T.int64(3)])
-                    for _k in range(T.int64(1)):
-                        out_buf[(fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) // T.int64(3) // T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) // T.int64(3) % T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) % T.int64(3)] = out_buf[(fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1) - T.int64(1)) * T.int64(3)) // T.int64(3) // T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1) - T.int64(1)) * T.int64(3)) // T.int64(3) % T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1) - T.int64(1)) * T.int64(3)) % T.int64(3)] + T.Cast("int32", rxplaceholder[(fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) // T.int64(3) // T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) // T.int64(3) % T.int64(2), (fused // T.int64(3) * T.int64(2) * T.int64(3) + fused % T.int64(3) + (_k + T.int64(1)) * T.int64(3)) % T.int64(3)])
-
-        @R.function
-        def main(x: R.Tensor((3, 2, 3), dtype="float32")) -> R.Tensor((3, 2, 3), dtype="int32"):
-            cls = Expected
-            gv = R.call_tir(cls.cumsum, (x,), out_sinfo=R.Tensor((3, 2, 3), dtype="int32"))
-            return gv
-    # fmt: on
-
-    mod = LegalizeOps()(Cumsum)
-    tvm.ir.assert_structural_equal(mod, Expected)
-
-
-def test_cumsum_symbolic():
-    # fmt: off
-    @I.ir_module
-    class Cumsum:
-        @R.function
-        def main(x: R.Tensor(("a", "b", "c"), "float32")):
-            gv = R.cumsum(x, axis=1, dtype="int32")
-            return gv
-
-    @I.ir_module
-    class Expected:
-        @T.prim_func
-        def cumsum(var_rxplaceholder: T.handle, var_cumsum_generic: T.handle):
-            T.func_attr({"tir.noalias": True})
-            a, b, c = T.int64(), T.int64(), T.int64()
-            rxplaceholder = T.match_buffer(var_rxplaceholder, (a, b, c), offset_factor=1)
-            out_buf = T.match_buffer(var_cumsum_generic, (a, b, c), "int32")
-            with T.block("cumsum_generic"):
-                T.reads(rxplaceholder[T.int64(0):a, T.int64(0):b, T.int64(0):c])
-                T.writes(out_buf[T.int64(0):a, T.int64(0):b, T.int64(0):c])
-                for fused in T.parallel(a * c):
-                    out_buf[(fused // c * b * c + fused % c) // c // b, (fused // c * b * c + fused % c) // c % b, (fused // c * b * c + fused % c) % c] = T.Cast("int32", rxplaceholder[(fused // c * b * c + fused % c) // c // b, (fused // c * b * c + fused % c) // c % b, (fused // c * b * c + fused % c) % c])
-                    for _k in range(b - T.int64(1)):
-                        out_buf[(fused // c * b * c + fused % c + (_k + T.int64(1)) * c) // c // b, (fused // c * b * c + fused % c + (_k + T.int64(1)) * c) // c % b, (fused // c * b * c + fused % c + (_k + T.int64(1)) * c) % c] = out_buf[(fused // c * b * c + fused % c + (_k + T.int64(1) - T.int64(1)) * c) // c // b, (fused // c * b * c + fused % c + (_k + T.int64(1) - T.int64(1)) * c) // c % b, (fused // c * b * c + fused % c + (_k + T.int64(1) - T.int64(1)) * c) % c] + T.Cast("int32", rxplaceholder[(fused // c * b * c + fused % c + (_k + T.int64(1)) * c) // c // b, (fused // c * b * c + fused % c + (_k + T.int64(1)) * c) // c % b, (fused // c * b * c + fused % c + (_k + T.int64(1)) * c) % c])
-
-        @R.function
-        def main(x: R.Tensor(("a", "b", "c"), dtype="float32")) -> R.Tensor(("a", "b", "c"), dtype="int32"):
-            a = T.int64()
-            b = T.int64()
-            c = T.int64()
-            cls = Expected
-            gv = R.call_tir(cls.cumsum, (x,), out_sinfo=R.Tensor((a, b, c), dtype="int32"))
-            return gv
-    # fmt: on
-
-    mod = LegalizeOps()(Cumsum)
-    tvm.ir.assert_structural_equal(mod, Expected)
-
-
-def test_einsum():
-    # fmt: off
-    @I.ir_module
-    class Einsum:
-        @R.function
-        def main(x: R.Tensor((2, 3), "float32"), y: R.Tensor((3, 4), "float32")):
-            gv = R.einsum((x, y), subscripts="ij,jk->ik")
-            return gv
-
-    @I.ir_module
-    class Expected:
-        @R.function
-        def main(
-            x: R.Tensor((2, 3), dtype="float32"), y: R.Tensor((3, 4), dtype="float32")
-        ) -> R.Tensor((2, 4), dtype="float32"):
-            cls = Expected
-            gv = R.call_tir(cls.einsum, (x, y), out_sinfo=R.Tensor((2, 4), dtype="float32"))
-            return gv
-
-        @T.prim_func
-        def einsum(
-            rxplaceholder: T.Buffer((T.int64(2), T.int64(3)), "float32"),
-            rxplaceholder_1: T.Buffer((T.int64(3), T.int64(4)), "float32"),
-            T_einsum: T.Buffer((T.int64(2), T.int64(4)), "float32"),
-        ):
-            T.func_attr({"tir.noalias": T.bool(True)})
-            for ax0, ax1, j in T.grid(T.int64(2), T.int64(4), T.int64(3)):
-                with T.block("T_einsum"):
-                    v_ax0, v_ax1, v_j = T.axis.remap("SSR", [ax0, ax1, j])
-                    T.reads(rxplaceholder[v_ax0, v_j], rxplaceholder_1[v_j, v_ax1])
-                    T.writes(T_einsum[v_ax0, v_ax1])
-                    with T.init():
-                        T_einsum[v_ax0, v_ax1] = T.float32(0)
-                    T_einsum[v_ax0, v_ax1] = (
-                        T_einsum[v_ax0, v_ax1]
-                        + rxplaceholder[v_ax0, v_j] * rxplaceholder_1[v_j, v_ax1]
-                    )
-    # fmt: on
-
-    mod = LegalizeOps()(Einsum)
-    tvm.ir.assert_structural_equal(mod, Expected)
-
-
-def test_einsum_symbolic():
-    # fmt: off
-    @I.ir_module
-    class Einsum:
-        @R.function
-        def main(x: R.Tensor(("a", "b"), "float32"), y: R.Tensor(("b", "c"), "float32")):
-            gv = R.einsum((x, y), subscripts="ij,jk->ik")
-            return gv
-
-    @I.ir_module
-    class Expected:
-        @R.function
-        def main(
-            x: R.Tensor(("a", "b"), dtype="float32"),
-            y: R.Tensor(("b", "c"), dtype="float32"),
-        ) -> R.Tensor(("a", "c"), dtype="float32"):
-            a = T.int64()
-            c = T.int64()
-            b = T.int64()
-            cls = Expected
-            gv = R.call_tir(cls.einsum, (x, y), out_sinfo=R.Tensor((a, c), dtype="float32"))
-            return gv
-
-        @T.prim_func
-        def einsum(
-            var_rxplaceholder: T.handle,
-            var_rxplaceholder_1: T.handle,
-            var_T_einsum: T.handle,
-        ):
-            T.func_attr({"tir.noalias": T.bool(True)})
-            a, b = T.int64(), T.int64()
-            rxplaceholder = T.match_buffer(var_rxplaceholder, (a, b))
-            c = T.int64()
-            rxplaceholder_1 = T.match_buffer(var_rxplaceholder_1, (b, c))
-            T_einsum = T.match_buffer(var_T_einsum, (a, c))
-            for ax0, ax1, j in T.grid(a, c, b):
-                with T.block("T_einsum"):
-                    v_ax0, v_ax1, v_j = T.axis.remap("SSR", [ax0, ax1, j])
-                    T.reads(rxplaceholder[v_ax0, v_j], rxplaceholder_1[v_j, v_ax1])
-                    T.writes(T_einsum[v_ax0, v_ax1])
-                    with T.init():
-                        T_einsum[v_ax0, v_ax1] = T.float32(0)
-                    T_einsum[v_ax0, v_ax1] = (
-                        T_einsum[v_ax0, v_ax1]
-                        + rxplaceholder[v_ax0, v_j] * rxplaceholder_1[v_j, v_ax1]
-                    )
-    # fmt: on
-
-    mod = LegalizeOps()(Einsum)
-    tvm.ir.assert_structural_equal(mod, Expected)
-
-
 def test_flip():
     # fmt: off
     @I.ir_module
