+++ tests/python/relax/test_tvmscript_parser_op_statistical.py
@@ -170,5 +170,20 @@ def test_std():
     _check(foo, bb.get()["foo"])
 
 
+def test_cumsum():
+    @R.function
+    def foo(x: R.Tensor((2, 3, 4), "float32")):
+        gv = R.cumsum(x, axis=1, dtype="int32")
+        return gv
+
+    x = relax.Var("x", R.Tensor((2, 3, 4), "float32"))
+    bb = relax.BlockBuilder()
+    with bb.function("foo", [x]):
+        gv = bb.emit(relax.op.cumsum(x, axis=1, dtype="int32"))
+        bb.emit_func_output(gv)
+
+    _check(foo, bb.get()["foo"])
+
+
 if __name__ == "__main__":
     tvm.testing.main()
